+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- Fixed an error that could occur when installing Commerce with an existing project config. ([#3472](https://github.com/craftcms/commerce/issues/3472))
+- Added `craft\commerce\helpers\ProjectConfigData::ensureAllStoresProcessed()`.
+
 ## 5.0.2 - 2024-05-01
 
 - Fixed a bug where setting a default tax zone would unset the default zone for all other stores. ([#3473](https://github.com/craftcms/commerce/issues/3473))